+++ src/main/java/com/getindata/connectors/http/internal/config/HttpConnectorConfigConstants.java
@@ -88,6 +88,18 @@ public final class HttpConnectorConfigConstants {
     public static final String SOURCE_CONNECTION_TIMEOUT =
         SOURCE_LOOKUP_PREFIX + "connection.timeout";
 
+    public static final String SOURCE_PROXY_HOST =
+            SOURCE_LOOKUP_PREFIX + "proxy.host";
+
+    public static final String SOURCE_PROXY_PORT =
+            SOURCE_LOOKUP_PREFIX + "proxy.port";
+
+    public static final String SOURCE_PROXY_USERNAME =
+            SOURCE_LOOKUP_PREFIX + "proxy.username";
+
+    public static final String SOURCE_PROXY_PASSWORD =
+            SOURCE_LOOKUP_PREFIX + "proxy.password";
+
     public static final String SINK_HTTP_TIMEOUT_SECONDS =
         GID_CONNECTOR_HTTP + "sink.request.timeout";