+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+-   Added option to define a proxy for the lookup source (including authentication)
+
 -   Added support for generic json and URL query creator 
 
 -   Retries support for source table: